Todo App

+++ Todo Server/src/main/java/com/ubedpathan/TodoApp/service/TodoService.java
@@ -1,9 +1,13 @@
 package com.ubedpathan.TodoApp.service;
 
 import com.ubedpathan.TodoApp.entity.TodoEntries;
+import com.ubedpathan.TodoApp.entity.UserEntity;
 import com.ubedpathan.TodoApp.repository.TodoRepository;
+import com.ubedpathan.TodoApp.repository.UserRepository;
+import lombok.extern.slf4j.Slf4j;
 import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.userdetails.User;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
@@ -13,23 +17,43 @@ import java.util.Map;
 import java.util.Optional;
 
 @Service
+@Slf4j
 public class TodoService {
     List<TodoEntries> todos = new ArrayList<>();
 
     @Autowired
     private TodoRepository todoRepository;
 
+    @Autowired
+    private UserRepository userRepository;
 
-    public String addTodos(TodoEntries todoEntries) {
-            todoEntries.setDate(LocalDateTime.now());
-            todoRepository.save(todoEntries);
-            return "Added successfully !";
+    @Autowired UserService userService;
+
+
+    public String addTodos(TodoEntries todoEntries, String userName) {
+
+            try{
+                UserEntity userData = userRepository.findByUsername(userName);
+                todoEntries.setDate(LocalDateTime.now());
+                TodoEntries saved = todoRepository.save(todoEntries);
+                userData.getTodoEntriesList().add(saved);
+                userService.saveUserTodoEntry(userData);
+                return "Added successfully !";
+            }catch (Exception e){
+                log.error("Error",e);
+                return null;
+            }
     }
 
-    public List<TodoEntries> getAllTodos() {
-        List<TodoEntries> allTodos = todoRepository.findByCompletedFalse();
-        return allTodos;
+    public List<TodoEntries> getUserTodos(String userName) {
+        List<TodoEntries> userTodos = userService.getUserTodos(userName);
+        if(userTodos != null) {
+            return userTodos;
+        }else{
+            return new ArrayList<>();
+        }
     }
+
 
     public boolean updateTodo(Map<String, Object> editTodo) {
         ObjectId todoId = new ObjectId(String.valueOf(editTodo.get("id")));
@@ -47,14 +71,24 @@ public class TodoService {
 
     }
 
-    public boolean deleteTodoById(ObjectId id) {
-        Optional<TodoEntries> findTodoById = todoRepository.findById(id);
-        if(findTodoById.isEmpty()){
-            return false;
+    public boolean deleteTodoById(ObjectId id, String userName) {
+        boolean removed = false;
+        UserEntity userEntity = userService.getUserByUserName(userName);
+        removed = userEntity.getTodoEntriesList().removeIf(x -> x.getId().equals(id));
+        if(removed){
+            userService.saveUserTodoEntry(userEntity);
+            Optional<TodoEntries> findTodoById = todoRepository.findById(id);
+            if(findTodoById.isEmpty()){
+                return false;
+            }
+            else {
+                todoRepository.deleteById(id);
+                return true;
+            }
         }
         else{
-            todoRepository.deleteById(id);
-            return true;
+            return false;
         }
+
     }
 }
